#888 Fix large net message packet fragmentation (#892)
* #888 root cause fix.

* #888 ERROR_DROP effectively zombies a dedicated server. Don't do that.

* #888 In fact, just don't overflow, ever.

If someone blows through the mem_buf_t, fully shit pants instead of pretending we'll fix it.

* Fix unrelated compiler warning.

* Fix some compiler warnings.

* Move the error into Mem_AllocBuffer.

* Fix format string.

+++ src/cgame/default/cg_entity.c
@@ -123,7 +123,7 @@ void Cg_LoadEntities(void) {
       if (!q_strcmp(classname, (*clazz)->classname)) {
 
         cg_entity_t e = {
-          .id = MAX_ENTITIES + cg_entities->count,
+          .id = MAX_ENTITIES + (int32_t) cg_entities->count,
           .clazz = *clazz,
           .def = def
         };
